feat(MovimientoStock): Se cambia la logica del stock, en vez de ser un valor propio de productos se crea una tabla movimientosStock que guarda cada movimiento de cada producto

+++ Comercio.Domain/Entidades/DetalleVenta.cs
@@ -7,7 +7,7 @@ namespace Comercio.Domain.Entidades
         public int Id { get; set; }
         public int IdVenta { get; set; }
         public int IdProducto { get; set; }
-        public decimal Cantidad { get; set; }
+        public int Cantidad { get; set; }
         public decimal PrecioUnitario { get; set; }
         public decimal Subtotal { get; set; }
     }
